requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
gunicorn==21.2.0
whitenoise==6.6.0
boto3==1.34.162
django-storages==1.14.4